@@ -8,12 +8,29 @@
     "authors": [
         {
             "name": "terminal42 gmbh",
-            "homepage": "https://terminal42.ch"
+            "homepage": "https://www.terminal42.ch"
         }
     ],
+    "funding": [
+        {
+            "type": "github",
+            "url": "https://github.com/terminal42"
+        },
+        {
+            "type": "other",
+            "url": "https://ko-fi.com/terminal42"
+        }
+    ],
+    "support": {
+        "issues": "https://github.com/terminal42/swissbilling-api/issues",
+        "source": "https://github.com/terminal42/swissbilling-api"
+    },
     "require": {
-        "php": "^7.1 || ^8.0",
+        "php": "^8.1",
         "ext-soap": "*"
+    },
+    "require-dev": {
+        "terminal42/contao-build-tools": "dev-main"
     },
     "autoload": {
         "psr-4": {
