@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Terminal42\SwissbillingApi\Soap;
 
 use Terminal42\SwissbillingApi\Exception\SoapException;
@@ -10,22 +12,22 @@ abstract class Api extends \SoapClient
     /**
      * @throws SoapException
      */
-    public function __call($function_name, $arguments)
+    public function __call($name, $args): mixed
     {
-        foreach ($arguments as &$argument) {
-            $this->convertArgument($argument);
+        foreach ($args as &$arg) {
+            $this->convertArgument($arg);
         }
 
-        unset($argument);
+        unset($arg);
 
         try {
-            return parent::__call($function_name, $arguments);
+            return parent::__call($name, $args);
         } catch (\SoapFault $e) {
             throw new SoapException($e, $this);
         }
     }
 
-    private function convertArgument(&$value): void
+    private function convertArgument(mixed &$value): void
     {
         switch (true) {
             /** @noinspection PhpMissingBreakStatementInspection */
@@ -34,10 +36,10 @@ abstract class Api extends \SoapClient
                 // no break;
 
             case \is_array($value):
-                array_walk($value, [$this, 'convertArgument']);
+                array_walk($value, $this->convertArgument(...));
                 break;
 
-            case \is_object($value) && \method_exists($value, '__toString'):
+            case \is_object($value) && method_exists($value, '__toString'):
                 $value = (string) $value;
                 break;
         }
