@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Terminal42\SwissbillingApi\Soap;
 
 use Terminal42\SwissbillingApi\Type\CheckResult;
@@ -12,21 +14,18 @@ use Terminal42\SwissbillingApi\Type\Transaction;
 use Terminal42\SwissbillingApi\Type\TransactionStatus;
 
 /**
- * @method TransactionStatus EshopTransactionRequest(Merchant $merchant, Transaction $transaction, Debtor $debtor, integer $count, InvoiceItem[] $items)
- * @method TransactionStatus EshopTransactionDirect(Merchant $merchant, Transaction $transaction, Debtor $debtor, integer $count, InvoiceItem[] $items)
- * @method TransactionStatus EshopTransactionByPass(Merchant $merchant, Transaction $transaction, Debtor $debtor, integer $count, InvoiceItem[] $items)
+ * @method TransactionStatus EshopTransactionRequest(Merchant $merchant, Transaction $transaction, Debtor $debtor, integer $count, array<InvoiceItem> $items)
+ * @method TransactionStatus EshopTransactionDirect(Merchant $merchant, Transaction $transaction, Debtor $debtor, integer $count, array<InvoiceItem> $items)
+ * @method TransactionStatus EshopTransactionByPass(Merchant $merchant, Transaction $transaction, Debtor $debtor, integer $count, array<InvoiceItem> $items)
  * @method TransactionStatus EshopTransactionCancel(Merchant $merchant, string $transaction_ref, DateTime $order_timestamp)
  * @method TransactionStatus EshopTransactionAcknowledge(Merchant $merchant, string $transaction_ref, DateTime $order_timestamp)
  * @method TransactionStatus EshopTransactionConfirmation(Merchant $merchant, string $transaction_ref, DateTime $order_timestamp)
  * @method TransactionStatus EshopTransactionStatusRequest(Merchant $merchant, string $transaction_ref, DateTime $order_timestamp)
  * @method TransactionStatus EshopCreditNotification(Merchant $request, string $transaction_ref, DateTime $order_timestamp, float $amount)
- * @method CheckResult EshopTransactionCheck(Merchant $request, DateTime $startdate, DateTime $enddate, string $status)
+ * @method CheckResult       EshopTransactionCheck(Merchant $request, DateTime $startdate, DateTime $enddate, string $status)
  */
 class ApiV2 extends Api
 {
-    /**
-     * @inheritDoc
-     */
     public function __construct(bool $production)
     {
         $wsdl = 'https://secure.safebill.ch/EShopRequestV2StdSec.wsdl';
