@@ -1,10 +1,12 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Terminal42\SwissbillingApi\Type;
 
 /**
  * @property TransactionStatus $methodstatus
- * @property OrderItem[]       $order_table
+ * @property array<OrderItem>  $order_table
  */
 class CheckResult extends AbstractType
 {
