@@ -1,11 +1,13 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Terminal42\SwissbillingApi\Type;
 
 /**
- * @property string  $failure_text
- * @property string  $Invoice
- * @property integer $Length
+ * @property string $failure_text
+ * @property string $Invoice
+ * @property int    $Length
  */
 class InvoiceStatus extends AbstractType
 {
